+++ src/main/java/bio/terra/app/configuration/ECMConfiguration.java
@@ -10,6 +10,7 @@ import org.springframework.context.annotation.Configuration;
 public class ECMConfiguration {
 
   private String basePath;
+  private String rasIssuer;
 
   public String getBasePath() {
     return basePath;
@@ -17,5 +18,13 @@ public class ECMConfiguration {
 
   public void setBasePath(String basePath) {
     this.basePath = basePath;
+  }
+
+  public String getRasIssuer() {
+    return rasIssuer;
+  }
+
+  public void setRasIssuer(String rasIssuer) {
+    this.rasIssuer = rasIssuer;
   }
 }
