+++ src/main/resources/application.properties
@@ -117,8 +117,8 @@ rbs.poolId=datarepo_v1
 rbs.instanceUrl=https://buffer.tools.integ.envs.broadinstitute.org
 rbs.clientCredentialFilePath=/tmp/buffer-client-sa-account.json
 ecm.basePath=https://externalcreds.dsde-dev.broadinstitute.org
+ecm.rasIssuer="https://stsstg.nih.gov"
 usermetrics.appId=datarepo
 usermetrics.bardBasePath=
 usermetrics.metricsReportingPoolSize=10
 usermetrics.ignorePaths=/api/repository/v1/jobs/*,/api/repository/v1/configs*,/api/repository/v1/upgrade,/actuator/*,/configuration,/status
-drs.rasIssuer="https://stsstg.nih.gov"
